+++ src/main/java/com/adven/concordion/extensions/exam/entities/AbstractEntity.java
@@ -1,4 +1,4 @@
-package com.adven.concordion.extensions.exam.kafka;
+package com.adven.concordion.extensions.exam.entities;
 
 import lombok.NonNull;
