+++ src/main/java/com/adven/concordion/extensions/exam/entities/Entity.java
@@ -1,4 +1,4 @@
-package com.adven.concordion.extensions.exam.kafka;
+package com.adven.concordion.extensions.exam.entities;
 
 public interface Entity {
