+++ src/main/java/com/adven/concordion/extensions/exam/kafka/DefaultEventProcessor.java
@@ -1,5 +1,6 @@
 package com.adven.concordion.extensions.exam.kafka;
 
+import com.adven.concordion.extensions.exam.entities.Entity;
 import com.adven.concordion.extensions.exam.kafka.check.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
