+++ src/main/java/com/adven/concordion/extensions/exam/kafka/EmptyEntity.java
@@ -1,5 +1,6 @@
 package com.adven.concordion.extensions.exam.kafka;
 
+import com.adven.concordion.extensions.exam.entities.Entity;
 import lombok.EqualsAndHashCode;
 
 @EqualsAndHashCode
