+++ src/main/java/com/adven/concordion/extensions/exam/kafka/EventProcessor.java
@@ -1,5 +1,7 @@
 package com.adven.concordion.extensions.exam.kafka;
 
+import com.adven.concordion.extensions.exam.entities.Entity;
+
 public interface EventProcessor {
 
     boolean check(final Event<? extends Entity> eventToCheck, final boolean isAsync);
