+++ src/main/java/com/adven/concordion/extensions/exam/kafka/HeaderBlockParser.java
@@ -1,6 +1,7 @@
 package com.adven.concordion.extensions.exam.kafka;
 
 import com.adven.concordion.extensions.exam.html.Html;
+import com.adven.concordion.extensions.exam.html.HtmlBlockParser;
 import com.google.common.base.Optional;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
