+++ src/test/java/com/adven/concordion/extensions/exam/entities/AbstractEntityTest.java
@@ -1,4 +1,4 @@
-package com.adven.concordion.extensions.exam.kafka;
+package com.adven.concordion.extensions.exam.entities;
 
 import org.junit.Test;
