+++ src/test/java/com/adven/concordion/extensions/exam/entities/MockEntity.java
@@ -1,4 +1,4 @@
-package com.adven.concordion.extensions.exam.kafka;
+package com.adven.concordion.extensions.exam.entities;
 
 public final class MockEntity extends AbstractEntity {
