+++ src/test/java/com/adven/concordion/extensions/exam/entities/ProtoEntityTest.java
@@ -1,12 +1,12 @@
-package com.adven.concordion.extensions.exam.kafka.protobuf;
+package com.adven.concordion.extensions.exam.entities;
 
+import com.adven.concordion.extensions.exam.kafka.protobuf.TestEntity;
 import org.junit.Test;
 
 import static com.adven.concordion.extensions.exam.RandomUtils.anyInt;
 import static com.adven.concordion.extensions.exam.RandomUtils.anyString;
 import static com.adven.concordion.extensions.exam.kafka.EventUtils.goodClass;
 import static com.adven.concordion.extensions.exam.kafka.EventUtils.goodMessage;
-import static com.adven.concordion.extensions.exam.kafka.protobuf.TestEntity.Entity;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class ProtoEntityTest {
@@ -27,7 +27,7 @@ public class ProtoEntityTest {
     public void testIsEqualTo() {
         final String name = anyString();
         final int number = anyInt();
-        final Entity entity = Entity.newBuilder()
+        final TestEntity.Entity entity = TestEntity.Entity.newBuilder()
             .setName(name)
             .setNumber(number)
             .build();
@@ -40,7 +40,7 @@ public class ProtoEntityTest {
     @Test
     public void testIsNotEqualToWithBadJsonValue() {
         final ProtoEntity protoEntity = new ProtoEntity(anyString(), goodClass().getName());
-        final Entity entity = Entity.newBuilder()
+        final TestEntity.Entity entity = TestEntity.Entity.newBuilder()
             .setName(anyString())
             .setNumber(anyInt())
             .build();
@@ -51,7 +51,7 @@ public class ProtoEntityTest {
     @Test
     public void testIsNotEqualTo() {
         final ProtoEntity protoEntity = new ProtoEntity(goodMessage(), goodClass().getName());
-        final Entity entity = Entity.newBuilder()
+        final TestEntity.Entity entity = TestEntity.Entity.newBuilder()
             .setName(anyString())
             .setNumber(anyInt())
             .build();
