+++ src/test/java/com/adven/concordion/extensions/exam/kafka/EventBlockParserTest.java
@@ -1,5 +1,6 @@
 package com.adven.concordion.extensions.exam.kafka;
 
+import com.adven.concordion.extensions.exam.entities.Entity;
 import com.adven.concordion.extensions.exam.html.Html;
 import com.google.common.base.Optional;
 import org.junit.Before;
