+++ src/test/java/com/adven/concordion/extensions/exam/kafka/EventUtils.java
@@ -1,6 +1,6 @@
 package com.adven.concordion.extensions.exam.kafka;
 
-import com.adven.concordion.extensions.exam.kafka.protobuf.ProtoEntity;
+import com.adven.concordion.extensions.exam.entities.ProtoEntity;
 import com.adven.concordion.extensions.exam.kafka.protobuf.TestEntity;
 
 import static com.adven.concordion.extensions.exam.RandomUtils.anyInt;
