+++ src/test/java/com/adven/concordion/extensions/exam/kafka/check/AsyncMockTestFail.java
@@ -1,7 +1,7 @@
 package com.adven.concordion.extensions.exam.kafka.check;
 
 import com.adven.concordion.extensions.exam.kafka.Event;
-import com.adven.concordion.extensions.exam.kafka.protobuf.ProtoEntity;
+import com.adven.concordion.extensions.exam.entities.ProtoEntity;
 import com.adven.concordion.extensions.exam.kafka.protobuf.TestEntity;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.common.utils.Bytes;
