+++ src/test/java/com/adven/concordion/extensions/exam/kafka/check/SyncMockTest.java
@@ -3,7 +3,7 @@ package com.adven.concordion.extensions.exam.kafka.check;
 import com.adven.concordion.extensions.exam.kafka.DummyEventConsumer;
 import com.adven.concordion.extensions.exam.kafka.Event;
 import com.adven.concordion.extensions.exam.kafka.check.verify.MockVerifier;
-import com.adven.concordion.extensions.exam.kafka.protobuf.ProtoEntity;
+import com.adven.concordion.extensions.exam.entities.ProtoEntity;
 import org.apache.kafka.common.utils.Bytes;
 import org.junit.Before;
 import org.junit.Test;
