+++ src/test/java/com/adven/concordion/extensions/exam/kafka/check/WithReplyTest.java
@@ -3,7 +3,7 @@ package com.adven.concordion.extensions.exam.kafka.check;
 import com.adven.concordion.extensions.exam.kafka.DummyEventProducer;
 import com.adven.concordion.extensions.exam.kafka.Event;
 import com.adven.concordion.extensions.exam.kafka.EventUtils;
-import com.adven.concordion.extensions.exam.kafka.protobuf.ProtoEntity;
+import com.adven.concordion.extensions.exam.entities.ProtoEntity;
 import org.junit.Before;
 import org.junit.Test;
