+++ src/test/java/com/adven/concordion/extensions/exam/kafka/check/verify/NullAwareVerifierTest.java
@@ -1,7 +1,7 @@
 package com.adven.concordion.extensions.exam.kafka.check.verify;
 
 import com.adven.concordion.extensions.exam.kafka.Event;
-import com.adven.concordion.extensions.exam.kafka.protobuf.ProtoEntity;
+import com.adven.concordion.extensions.exam.entities.ProtoEntity;
 import org.apache.kafka.common.utils.Bytes;
 import org.junit.Before;
 import org.junit.Test;
